It takes 8.5% less time to parse according to the benchmark, It uses ~52.7% less memory, and makes 16.8% less allocations.

Note, there are no breaking changes.

```
$ go test -count=10 -bench=^BenchmarkUserAgent$ -benchmem .
$ benchstat old.txt new.txt

name          old time/op    new time/op    delta
UserAgent-12     237µs ± 2%     217µs ± 2%   -8.52%  (p=0.000 n=10+10)

name          old alloc/op   new alloc/op   delta
UserAgent-12    89.8kB ± 0%    42.5kB ± 0%  -52.75%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
UserAgent-12     1.82k ± 0%     1.52k ± 0%  -16.83%  (p=0.000 n=10+10)
```

<details>

<summary>old.txt</summary>

```
goos: darwin
goarch: amd64
pkg: github.com/mileusna/useragent
cpu: Intel(R) Core(TM) i5-10600 CPU @ 3.30GHz
BenchmarkUserAgent-12    	    4748	    234376 ns/op	   89910 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    5067	    235759 ns/op	   89902 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    5061	    241559 ns/op	   89815 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    4959	    241981 ns/op	   89761 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    4848	    238471 ns/op	   89880 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    4965	    238081 ns/op	   89819 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    4884	    232825 ns/op	   89921 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    4891	    235413 ns/op	   89790 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    5052	    237770 ns/op	   89868 B/op	    1824 allocs/op
BenchmarkUserAgent-12    	    4998	    237143 ns/op	   89719 B/op	    1824 allocs/op
PASS
ok  	github.com/mileusna/useragent	12.890s
```

</details>

<details>

<summary>new.txt</summary>

```
goos: darwin
goarch: amd64
pkg: github.com/mileusna/useragent
cpu: Intel(R) Core(TM) i5-10600 CPU @ 3.30GHz
BenchmarkUserAgent-12    	    5184	    214837 ns/op	   42490 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5476	    214017 ns/op	   42445 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5276	    214757 ns/op	   42505 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5454	    218415 ns/op	   42433 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5116	    217853 ns/op	   42408 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5482	    220742 ns/op	   42479 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5217	    218494 ns/op	   42431 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5294	    216651 ns/op	   42413 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5479	    218571 ns/op	   42472 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5474	    216790 ns/op	   42452 B/op	    1517 allocs/op
PASS
ok  	github.com/mileusna/useragent	12.049s
```

</details>