@mileusna 2023-11-09
Hello, thank your contribution.

As I can see it is not safe to use it concurrently, and in web apps user agents are usually parsed concurrently, for example on each HTTP request? Therefore it does represents breaking changes in existing apps.