@marselester 2023-11-09
Hi! That's a good point, I have updated the code and benchmarks.

```
$ benchstat old.txt new.txt 
name          old time/op    new time/op    delta
UserAgent-12     237µs ± 2%     221µs ± 1%   -6.73%  (p=0.000 n=10+10)

name          old alloc/op   new alloc/op   delta
UserAgent-12    89.8kB ± 0%    42.5kB ± 0%  -52.66%  (p=0.000 n=10+10)

name          old allocs/op  new allocs/op  delta
UserAgent-12     1.82k ± 0%     1.52k ± 0%  -16.83%  (p=0.000 n=10+10)
```

<details>

<summary>new.txt</summary>

```
goos: darwin
goarch: amd64
pkg: github.com/mileusna/useragent
cpu: Intel(R) Core(TM) i5-10600 CPU @ 3.30GHz
BenchmarkUserAgent-12    	    4928	    220713 ns/op	   42575 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5343	    221977 ns/op	   42474 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5191	    221726 ns/op	   42533 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5101	    223259 ns/op	   42576 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5205	    222943 ns/op	   42518 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5353	    221157 ns/op	   42544 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5388	    221315 ns/op	   42494 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5380	    220784 ns/op	   42487 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5241	    220295 ns/op	   42564 B/op	    1517 allocs/op
BenchmarkUserAgent-12    	    5188	    219393 ns/op	   42513 B/op	    1517 allocs/op
PASS
ok  	github.com/mileusna/useragent	12.211s
```

</details>