@marselester 2024-02-26
I understand your concerns. Thank you for considering this change.

The current version of the PR is safe to use concurrently. The [sync.Pool](https://pkg.go.dev/sync#Pool) handles all the synchronizations. Initially I wasn't sure if concurrency safety was part of the library API or left to the caller to handle.

I tested user agent parsing on ~600K headers with the following results: it took ~2s to process, consuming ~500 MB of memory, and allocating ~11.26M objects on the heap.

Extrapolating from the benchmarks, there could be 263.3 MB of memory saved, and 1.89 million less allocations which should reduce GC pressure since there will be less heap objects to check (shorter GC pauses, less CPU consumption, improved tail latency of HTTP requests if user agents are parsed while serving requests).

I have a follow up PR https://github.com/marselester/useragent/pull/1 that could further reduce memory consumption by -36.23% (~95.4 MB) and allocations by -9.43% (~0.88M).

So the same 600K headers could be processed using 141.3 MB of memory (vs 500 MB), and 8.49M of allocs (vs 11.26M). There might be more opportunities to improve performance.